Title: 최소 에너지 비용, Time: 13ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240403/최소 에너지 비용/minimum-energy-cost.cpp
@@ -32,7 +32,7 @@ int main() {
     }
     long long answer = 0;
     for(int i = 1; i < n; i++){
-        answer += move_cost[i] * charge_cost[min_charge_number[i]];
+        answer += (long long)move_cost[i] * charge_cost[min_charge_number[i]];
     }
 
     cout << answer;
